refactor: clean up cycles.rs — replace hand-rolled glob, extract TOON formatting, reuse ignore crate
1. Removed hand-rolled glob_match() in favor of the `ignore` crate's
   Gitignore matcher — same API used by rpg-encoder's evolution.rs for
   .rpgignore handling. Added `ignore` as a dependency to rpg-nav.

2. Extracted ~130 lines of inline TOON formatting from the detect_cycles
   tool handler in tools.rs into format_cycle_report() in toon.rs,
   matching the existing format_health_report() pattern.

3. Replaced CycleConfig's project_root + ignore_rpgignore fields with a
   single excluded_paths: Option<Gitignore>, moving .rpgignore loading
   responsibility to the MCP layer where it belongs.

Closes #62
